AlpacaServer/requirements.txt
django
pyserial
skyfield
redis
numpy